requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
qdrant-client==1.7.0
sentence-transformers==2.7.0
google-generativeai==0.3.2
pymupdf==1.23.8
python-multipart==0.0.6
pydantic==2.5.0
huggingface-hub==0.23.0